move username entry field to its own component
