+++ src/components/App/AppContainer.js
@@ -1,11 +1,6 @@
 import {connect} from 'react-redux';
 import App from './App';
-import {
-	setSlackUserFetchError,
-	setCurrentUser,
-	setPrediction,
-	loadUsers
-} from './actions';
+import {setSlackUserFetchError, setPrediction, loadUsers} from './actions';
 import {loadStoreList, setShowList} from './../StoreList/actions';
 
 const mapStateToProps = state => {
@@ -13,7 +8,6 @@ const mapStateToProps = state => {
 		storeList: state.storeList,
 		showList: state.showList,
 		slackUserFetchError: state.slackUserFetchError,
-		currentUser: state.currentUser,
 		users: state.users,
 		prediction: state.prediction
 	};
@@ -23,7 +17,6 @@ const mapDispatchToProps = dispatch => {
 	return {
 		loadUsers,
 		loadStoreList,
-		setCurrentUser: currentUser => dispatch(setCurrentUser(currentUser)),
 		setPrediction: prediction => dispatch(setPrediction(prediction)),
 		setShowList: showList => dispatch(setShowList(showList)),
 		setSlackUserFetchError: isError =>
