+++ src/components/App/actionTypes.js
@@ -1,4 +1,3 @@
 export const SET_USERS = 'SET_USERS';
 export const SET_SLACK_USER_FETCH_ERROR = 'SET_SLACK_USER_FETCH_ERROR';
-export const SET_CURRENT_USER = 'SET_CURRENT_USER';
 export const SET_PREDICTION = 'SET_PREDICTION';
