+++ src/components/App/actions.js
@@ -1,7 +1,6 @@
 import {
 	SET_USERS,
 	SET_SLACK_USER_FETCH_ERROR,
-	SET_CURRENT_USER,
 	SET_PREDICTION
 } from './actionTypes';
 
@@ -18,13 +17,6 @@ export function setSlackUserFetchError(slackUserFetchError) {
 	return {
 		type: SET_SLACK_USER_FETCH_ERROR,
 		slackUserFetchError
-	};
-}
-
-export function setCurrentUser(currentUser) {
-	return {
-		type: SET_CURRENT_USER,
-		currentUser
 	};
 }
