+++ src/components/UsernameEntry/actionTypes.js
@@ -1,0 +1,1 @@
+export const SET_CURRENT_USER = 'SET_CURRENT_USER';
