+++ src/components/UsernameEntry/actions.js
@@ -1,0 +1,8 @@
+import {SET_CURRENT_USER} from './actionTypes';
+
+export function setCurrentUser(currentUser) {
+	return {
+		type: SET_CURRENT_USER,
+		currentUser
+	};
+}
