+++ src/components/UsernameEntry/reducer.js
@@ -1,0 +1,6 @@
+import {SET_CURRENT_USER} from './actionTypes';
+
+export function currentUser(state = '', action) {
+	if (action.type === SET_CURRENT_USER) return action.currentUser;
+	return state;
+}
